Renamed special_tiles module to place_items -- better more apt name
