+++ src/map/generator/place_items.rs
